requirements.txt
python-dotenv==0.19.2
imutils==0.5.4
opencv-python==4.5.4.60
numpy==1.21.5